Migrate to DayJs for seconds parsing
